Removed image URL appending
Currently the parser is appending image urls to an existing image url if one exists. This leads to setting the field as http://example.com/image.pnghttp://example.com/image2.png if there are mutliple images listed for the feed. This change makes it so the image URL field gets replaced, instead of appended to, whenever it finds a new image.